refactor(DatabaseUtils) : 패키지 이동
