+++ src/main/java/janggi/board/dao/JanggiBoardDAOImpl.java
@@ -2,7 +2,7 @@ package janggi.board.dao;
 
 import janggi.board.JanggiBoard;
 import janggi.board.PieceFactory;
-import janggi.board.dao.utils.DatabaseUtils;
+import janggi.database.utils.DatabaseUtils;
 import janggi.piece.Piece;
 import janggi.piece.PieceType;
 import janggi.value.JanggiPosition;
