+++ src/test/java/janggi/board/dao/TeamDAOImplTest.java
@@ -1,6 +1,6 @@
 package janggi.board.dao;
 
-import janggi.board.dao.utils.DatabaseUtils;
+import janggi.database.utils.DatabaseUtils;
 import janggi.database.DBConnectorTest;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
